Adicionado metodo para deletar arquivo depois dos testes

+++ src/ProcessaArquivosTest.java
@@ -4,14 +4,17 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+
+import org.junit.After;
+import org.junit.Before;
 import org.junit.jupiter.api.Test;
 
 class ProcessaArquivosTest {
 
-	@Test
-	void listaArquivos() {
+	@Before
+	void criaArquivoParaTeste() {
 		String nomeArquivo = "01.txt";
-		
+
 		File file = new File("/home/maicon/eclipse-workspace/Procesador de Arquivos/data/in/" + nomeArquivo);
 		boolean aquivoExiste = file.exists();
 		if (!aquivoExiste) {
@@ -21,30 +24,32 @@ class ProcessaArquivosTest {
 				io.printStackTrace();
 			}
 		}
+	}
+
+	@Test
+	void listaArquivosAprocessar() {
+		String nomeArquivo = "01.txt";
 
 		String caminhoArquivo = "data/in/";
 		ProcessaArquivos processa = new ProcessaArquivos();
 		String conteudoDiretorio = processa.listaArquivos(caminhoArquivo);
 		String expected = nomeArquivo;
-		file.delete();
 		assertEquals(expected, conteudoDiretorio);
 	}
+
+	/*
+	 * @Test void listaArquivosProcessados() { String nomeArquivo =
+	 * "01-processado.txt"; String caminhoArquivo = "data/out/"; ProcessaArquivos
+	 * processa = new ProcessaArquivos(); String conteudoDiretorio =
+	 * processa.listaArquivos(caminhoArquivo); String expected = nomeArquivo;
+	 * assertEquals(expected, conteudoDiretorio); }
+	 */
 
 	@Test
 	void leArquivoPorNome() {
 		String nomeArquivo = "01.txt";
 		String caminhoArquivo = "data/in/";
-		
-		File file = new File("/home/maicon/eclipse-workspace/Procesador de Arquivos/data/in/" + nomeArquivo);
-		boolean aquivoExiste = file.exists();
-		
-		if (!aquivoExiste) {
-			try {
-				file.createNewFile();
-			} catch (IOException io) {
-				io.printStackTrace();
-			}
-		}
+
 		ProcessaArquivos processa = new ProcessaArquivos();
 		String conteudoArquivo = processa.leArquivoPorNome(caminhoArquivo, nomeArquivo);
 		String expected = "";
@@ -63,8 +68,19 @@ class ProcessaArquivosTest {
 		} catch (IOException e) {
 			System.err.printf("Erro na abertura do arquivo: %s.\n" + e.getMessage(), e.getMessage());
 		}
-		file.delete();
+
 		assertEquals(expected, conteudoArquivo);
+	}
+
+	@After
+	void removeArquivoParaTeste() {
+		String nomeArquivo = "01.txt";
+		File file = new File("/home/maicon/eclipse-workspace/Procesador de Arquivos/data/in/" + nomeArquivo);
+			
+		boolean aquivoExiste = file.exists();
+		if (aquivoExiste) {
+			file.delete();
+		}
 	}
 
 }
